Read/write from/to files and break up main.c
main.c now mainly does housekeeping stuff around argument parsing while
the main secret-sharing stuff is in libsss.

The decoding function now reads shares from files instead of stdin.

Also update README and makefile with libsss.c and fix target name.

If this was for work I'd split this commit up properly but eh, it's a
side project. I'm also not happy with the file-checking stuff being in
libsss but I'll just fix that later. It works so just commit it!