Upgrade to 256-bit secrets with GMP
To accommodate the larger size we switch to a new prime p.

Most changes are just switching to the gmp versions of the same
arithmetic. We also fix some input errors returning 0 instead of 1 and
the duplicated coordinates logic.

f_norm and finite_inverse are provided for us by GMP so we delete our
implementations.

Finally, update and fix typos in README.