Rewrite argument parsing with version info
The -v/--version option is added and -e/-d now work correctly, although
this does involve changing some default behaviour.  We also add
additional tests for some (not all) different styles of invocation.

The whole X_STR + STR thing is weird but seems to be the "officially"
sanctioned way of doing it in C so alright. Unless I'm mistaken this way
is better as it allows the combined version string to be built at
compile-time. I'm not 100% about the formatting and linebreaks etc. but
I'll go mad bikeshedding it so this is fine for now.

Also update the comment in sss.c which was already out of date already.