Rewrite file handling API and format
To keep the library code as focused as possible we require callers to
open files, allocate buffers and handle decoding errors themselves -
this is reflected in sss.c.

We introduce a new structure, sss_share, to represent a share. For now
it only contains the x and y coordinates but in future will include
things like the associated prime number and so on when these are no
longer fixed.

Writing shares to and from files is delegated to new (de)serialisation
functions, with a new versioned binary format.

Finally, update the header and comment for legibility.