Refactor Makefile with implicit rules
We remove the manual cc invocations and allow implicit rule chains to
build the sss executable. Most of the config now lives in the CFLAGS,
LDLIBS and OBJECTS variables.

To allow the implicit rules to work for debug builds we must use sss as
the executable name again. Therefore we set some flags as
target-specific variable values and use the lastflags target to check if
object files need to be rebuilt.

Finally, because of changes to the intermediate files we update the
clean target and gitignore.