Refactor decoding to use arrays
This is more prep for when we allow k > 2, although for now we just
hardcode everything to use 2 shares still.

Now that we deal with arrays I don't think we can return structs from
sss_des any more, so we convert it to a (probably more sensible) pointer
signature. Otherwise we just convert struct accesses in sss_rec and
sss_dec and put the share deserialisation calls into a loop.

Also remove "internal" functions from libsss.h for now, as it's tedious
to keep them in sync and also not relevant if you think of the header as
the public API. Not sure exactly what standard practice is though.