Allow generating shares with k = 3
We add k to the signature of sss_enc to allow setting the share recovery
threshold when generating shares. To allow this we convert a1 into an
array of coefficients and break out the coefficient generation + the y
coordinate calculation into loops.

Also convert *shares to shares[n] for consistency in sss_enc and update
some comments.

In sss.c we add a -k/--threshold option in the same vein as -n/--shares,
which we restrict to 2 or 3 for now. There's no reason sss_enc couldn't
generate shares with k > 3 but let's see how tricky it is to generalise
the decoding logic first.

Amusingly, shares generated with this code will only be recoverable by
future versions of the program as it currently can't decode when k = 3.