Allow any k >=2 in libsss when decoding shares
Instead of hardcoded arithmetic to calculate the secret we replace the
sss_rec function with a generic routine to evaluate the Lagrange
interpolating polynomial at zero instead. This allows decoding shares
for any k, although for now we only allow up to k=9 in sss to avoid
excessive rewriting.

Argument handling requires a rewrite to remove the assumption that only
two arguments will be of interest. Instead of handling the args before
the enc/dec block we process them differently in each block, with as few
extra variables as possible. This version isn't great and could be
improved but that can come later as it's more about enabling the new k
threshold for now.

Also change MAX_THRESHOLD in sss.c to 9 as it works so there's no point
in not doing it immediately.